memory: Simplify the memory allocator implementation
This (admittedly massive) path simplifies the way memory allocation is
performed in the library.

The old `gumbo_parser_allocate` APIs have been removed and replace with
the following:

- `gumbo_malloc`
- `gumbo_realloc`
- `gumbo_free`
- `gumbo_strdup`

As you can see, the 4 APIs match their C standard equivalents (in both
function and signature), and they no longer take a `GumboParser *`
object to lookup their implementation.

Instead, their implementation can be customized, globally, using the
following APIs:

- `gumbo_memory_set_allocator`
- `gumbo_memory_set_free`

These two APIs allow the user to set a global memory allocator and free
function. The `allocator` function needs to have the same signature as
the standard `realloc` (this allows us to use it both as a realloc in
the vector and string buffer code, *greatly* reducing memory usage), and
as a normal malloc (by passing `NULL` as the first argument).

The `free` function needs to have the same signature as the standard
`free`.

With just these two functions, we can abstract the whole set of standard
C memory allocation APIs, and we can do so globally, without having to
pass around the parser state to find them.

This greatly simplifies many parts of the library, improves performance,
and fixes several pathological cases of excessive memory usage,
caused by the previous lack of a `realloc` API.

The following external APIs, however, are no longer backwards
compatible:

- struct GumboInternalOptions: no longer allows the user to set a custom
   memory allocator callback.

- gumbo_destroy_output: no longer requires a Parser object.

- gumbo_destroy_node: can now be safely exported

Conflicts:
	src/parser.c
	src/tokenizer.c